Address review notes
- The observation imports in `viz.core` were function-local because
  `viz.observations` needed `percentile_norm` from `viz.core`, which is a cycle
  I introduced. Fixed at the root: the colour-limit helpers move to
  `viz.norms`, which neither imports, so every import here is top level again.
- The `observations:` include moves from the v2 preset to `samudra_om4`, for
  both `eval.yaml` and `viz.yaml`. That is the preset the README already runs
  and the one we iterate on.
- `Comparison.trimmed` takes a `time_slice`, matching `Om4TimeConfig`.
- cartopy, cmocean and tqdm become ordinary dependencies and the `viz` extra
  goes. They add 40 MB across 8 packages, all binary wheels, no compilation,
  and resolve on both aarch64 and x86_64 manylinux -- next to torch that is
  noise, and viz is a normal part of the workflow rather than an optional
  feature. The README's `uvx samudra viz` line then needs no change, which was
  the point of the note about it.
