+++ src/samudra/configs/samudra_om4/viz.yaml
@@ -22,3 +22,4 @@ groundtruth_time_range:
 runs:
   - name: "samudra_om4_eval"
     location: ".LOCAL/samudra_om4_eval/predictions.zarr"
+observations: !include ../data/obs.yaml
