+++ src/samudra/configs/samudra_om4_v2/eval.yaml
@@ -18,4 +18,3 @@ experiment:
     project: default
 data: !include ../data/om4.yaml
 model: !include model.yaml
-observations: !include ../data/obs.yaml
